[op][n_upsable][error_page] Remove a useless parameter related to LoadFailed
:Release Notes:
Remove a useless parameter which is always the empty string

:Detailed Notes:
Since Chromium68, the third argument of WebViewDelegate::LoadFailed has
been the empty string.
In the past versions, WebView::DidFinishNavigation called
WebView::DidFailLoad with base::ASCIIToUTF16(""), and
WebView::DidFailLoad forwarded base::UTF16ToUTF8(error_description) to
WebViewDelegate::LoadFailed.
Recently, WebView::DispatchNetError has called
WebViewDelegate::LoadFailed with std::string().
The value of the error description is always the empty string, and this
value seems useless in WAM. Specifically, the value of the error
description is just used to compare it with "SSL_ERROR" in
WebPageBlink::loadFailed.
However, the actual value of the error description is always "".
For these reasons, the parameter, const std::string& error_description,
would be removed.

There was a regression in CCC and this patch was reverted.
However, this patch does not have to do with the regression.

For these reasons, this patch would be applied again.

:Testing Performed:
Build OK

:QA Notes:
Test-cases: N/A
Depends on: http://gpro.lge.com/c/webosose/wam/+/390497

:Issues Addressed:
[WRQ-8302] [error_page] Update some commits related to an error page
[WRO-9713] [Error page] Investigate usage of SSL_ERROR in WAM
[WRO-12944] CCC: pro: webruntime_94=pro.27

Change-Id: Ib9995ddc1c768d686b08827d403678263969f91c
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/336818
Reviewed-on: http://gpro.lge.com/c/webosose/chromium108/+/390499
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>