fixup! [op][n_upsable][appshell] Fix the issue with the tooltip not displaying
:Release Notes:
Fix the issue with the tooltip not displaying

:Detailed Notes:
The problem is that the tooltip is not being shown.
This feature was not added in the app-shell side.
To solve this, we create a tooltip object
and assign it to the root window to listen for mouse move events.
Note that the ViewsDelegate interface is not implemented in the appshell
side, which leads to a crash issue when initializing tooltips.

:Testing Performed:
1. Run Enact browser Application
2. Open youtube link. https://youtube.com
3. Move mouse focus to "All" button
4. Tooltip will show under cursor

:QA Notes:
Test-cases: QWTC-12840

:Issues Addressed:
[WRP-25056] Port fixup! [op][n_upsable][appshell] Fix the issue
with the tooltip not displaying
[WRP-15770] Description of selectable tabs No speech bubble

Change-Id: I8fd09266161e77200b9fb1aea613aa5dd7104e6d
Originally-Reviewed-on:
http://gpro.lge.com/c/webosose/chromium94/+/357816
Reviewed-on: http://gpro.lge.com/c/webosose/chromium108/+/369593
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Tested-by: Van Luc Tran <luc2.tran@lge.com>