fixup! [op][n_upsable][luna] Fix LunaServiceClient for utility process
:Release Notes:
Fix LunaServiceClient to use g main content thread defult for luna
response

:Detailed Notes:
Currently in LunaServiceClient we use g_main_context_default for
utility process as well. But as a result we need to thread
switches in several places since in this approach we receive the
luna response in main thread where as many calls for audio and
camera service are also invoked in main thread. So the usage luna
service client for utility process becomes difficult.
So in this patch set we use use g_main_context_get_thread_default
for utility process. Removed the UI message pump type in utility
main. So that individual services can start the thread with UI
type message pump for invoking luna calls
As a result luna response will be received in the caller thread
and keep the main thread unblocked.

:Testing Performed:
Compilation is successfull. Audio and video service functionality
were checked with their respective fixup patch sets

:QA Notes:
Test-cases: QWTC-5929, QWTC-5924

:Issues Addressed:
[WRP-27775] Port fixup! [op][n_upsable][luna] Fix LunaServiceClient for utility process

Change-Id: Ia3f47eff6fbaf0268303a661ce7466fea56674b8
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/366134
Reviewed-on: http://gpro.lge.com/c/webosose/chromium108/+/371220
Tested-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>