[op][n_upsable][media] Fix punch hole with blending for opacity
:Release Notes:
Fix video visibility issue with blending for opacity < 1.0f

:Detailed Notes:
Now in this version of chromium when the control is hidden than
the opacity (if < 1.0f) value is getting applied to the video layer.
As a result if we apply transparent color with blend mode SrcOver
than the video is not visible because blend mode is not changed
to Src in SkiaRenderer::CalculateDrawQuadParams.
In order to preserve the opacity and make the video visible we
have to set the solid color Black with blend mode DstOut for this
scenario.

Referred this logic from ProcessForUnderlay of dc_layer_overlay.cc

:Testing Performed:
1. Launch a test app with below HTML content in test page
  <html>
  <body>
    <video width="640" height="480" style="opacity: 0.5" autoplay controls muted>
      <source src="https://commondatastorage.googleapis.com/gtv-videos-bucket/sample/BigBuckBunny.mp4">;
    </video>
  </body>
  </html>
2. Ensure video is shown and has shaded foreground layer

:QA Notes:
Test-cases: QWTC-4117

:Issues Addressed:
[WRP-11054] : Video is not visible properly

Change-Id: I9455e06ed54d1494c0ac1029ffbeab7ac7f1677d
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium108/+/351449
Reviewed-on: http://gpro.lge.com/c/webosose/chromium108/+/373018
Tested-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>