[oe][n_upsable][pwa] Update application resources
:Release Notes:
Update PWA resources (like icons) when they are changed on the site

:Detailed Notes:
Update happens in WAM when PWA page is loaded.
Use WebViewGuestInstallableManager in WebView and when app is
navigated to an url kick off the app update. The update goes
same path as install: downloads resources and forms AppInfo.
Then it calls InstallWebApp with the new AppInfo.
During the update there are many points where it cancelled, for these
external API should be implemented, e.g. not update when app is being
updated already, when it was updated not so long ago, and most
importantly when the web app manifest and its resources (icons etc.) are
not changed.

The new contents of the app is placed in the temporary directory and
then appinstallservice is called to perform actual app update.

The com.webos.app.home loads app icons through file:// protocol. It
is non-trivial how to make it reload updated image from disk when app is
updated. One way is to change the icon file name every time the new
version is installed. For this we add timepstamp numeral to the names of
the icons.

Depends on changes in com.webos.app.home:
http://gpro.lge.com/c/webosose/com.webos.app.home/+/371720

:Testing Performed:
1. Navigate browser to PWA enabled site.
2. Click the install icon in the toolbar.
EXP: the icon in the browser disappears.

3. Switch to Home
EXP: Newly installed app shown in the apps

4. Update the webapp in hosting website

5. click the PWA app
EXP: app is launched

6. Switch to Home
EXP: app icon is updated

:QA Notes:
Test-cases: QWTC-13576, QWTC-13585

:Issues Addressed:
[WRP-29170] Integrate and adapt the update solution

Change-Id: Ieef994c0f156c86f2d7a1adcc3479176accb4b2c
Reviewed-on: http://gpro.lge.com/c/webosose/chromium108/+/380739
Reviewed-by: Automated Checker <automated_checker@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Vladislav Mukulov <vladislav.mukulov@lge.com>
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>