[op][n_upstable][cookie] Send OSCrypt raw encryption key to Network Service
:Release Notes:
Send OSCrypt raw encryption key to Network Service

:Detailed Notes:
The chromium bug, https://crbug.com/1277463, was fixed in M99. It is
related to modifying os_crypt_linux.cc to allows getting and setting the
raw encryption key and the encryption key then sent to network service
to initialize its instance of os_crypt. In this patch, we also applied
this changes to neva.

Referring to
https://chromium-review.googlesource.com/c/chromium/src/+/3320484,
An os_crypt::Config should have been set before starting the network
service. So, It needs to initialize the config in the BrowserMainParts
of neva and remove some extra code for initialized os_crypt config.
For setting encryption key to network service for neva based on Linux,
it needs to call OSCrypt::GetRawEncryptionKey() in the browser process
and send it to the network service via
NetworkService::SetEncryptionKey().

:Testing Performed:
1. Change log level
 # vi /lib/systemd/system/scripts/webapp-mgr.sh
 edit WAM_EXTRA_FLAGS="--v=3"
2. Remove a directory for cookies
 # rm -rf /var/lib/wam/Default/Cookies
3. Restart wam
 # restart webapp-mgr
4. Print log by connecting another terminal
 # journalctl -u webapp-mgr.service -p "debug" -f | grep crypt
5. Launch any web app
6. Open web inspector and move the url by using console panel.
  location.href="https://titan.lge.com/blink-tests/web/cookie.html"
7. Print logs like "EncryptString"
8. Repeat step 3~6
9. Print logs like "DecryptString"

:QA Notes:
Test-cases: QWTC-6006, QWTC-7256, QWTC-7333

:Issues Addressed:
[WRP-29581] After implementation DecryptString logs are not listing
[WRP-29560] Cookie input is not updated

Change-Id: I3eabc711d98d02bce5c6ebb1c255f402800880f4
Reviewed-on: http://gpro.lge.com/c/webosose/chromium108/+/376323
Tested-by: Jeongran Jang <jeongran.jang@lge.com>
Reviewed-by: Jeongran Jang <jeongran.jang@lge.com>
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Joonhyung Hwang <joonhyung.hwang@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>