[op][n_upsable][appshell] Add injection API for Custom UserAgent feature
:Release Notes:
Add injection API for Custom UserAgent feature.

:Detailed Notes:
OSE is open source code, all information is public and sensitive
information is not hidden. Specifically, here is information about the
server (url and apikey). If this information is published, it will
lead to the risk of leaking security information as well as hackers
being able to penetrate the system.
To address this issue, we create files to store this information,
These files will be encrypted, and stored locally on the device.
And can pass to JS side via injectionAPI mechanism.
This is a feature for internal use only and is not public.

:Testing Performed:
1. Close Web Browser
2. Open Web inspector console for Enact browser
3. Create Encrypt Data:
  navigator.customuseragent.createEncryptedServerCredentials(
    "{\"api_auth_key\":\"UDBjSiZnMjAxX0AxQHUwaA==\",\"url\":
	\"http://10.218.140.120:12345/cua.json\"}",
	"customUA_0123456",(value) => {
        console.log(value)
    });
4. Close Web Browser
5. Making file custom_ua_auth_decryption_key and save "encrypt_key" to
this file. Move this file to /var/cache/xdg/app_shell- and
/var/cache/xdg/app_shell-0
6. Making file custom_ua_service_auth_data and save Encrypt data which
was create in step3 to this file. Move this file to
/var/cache/xdg/app_shell- and  /var/cache/xdg/app_shell-0
7. Close Web Browser
8. Open Web inspector console for Enact browser
9. Get Custom UserAgent Data
   navigator.customuseragent.getServerCredentials((value) => {
            console.log(value)
        });
10. Close Web Browser

:QA Notes:
Test-cases: QWTC-13638

:Issues Addressed:
[WRQ-944] Investigate the mechanism to hide the authentication key

Change-Id: I0ff17c73146fd349602449d3f8e6333c7f6e3f36
Reviewed-on: http://gpro.lge.com/c/webosose/chromium108/+/379634
Reviewed-by: Automated Checker <automated_checker@lge.com>
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>