[oe][n_upsable][gpu] Fix GL initialization for VMware
:Release Notes:
Fixed GL initialization for VMware

:Detailed Notes:
A new rule was added to software_rendering_list.json that blocks
the correct GL initialization for webos-image. This rule is based
on assumption that "VMware is buggy on Linux" and blocks gpu
initialization. As far as VMware works correctly for the webOS
needs, according to the tests, it is needed to re-enable the
correct gpu initialization for webOS VMware.

:Testing Performed:
1. Prepare webos-image-devel
2. Run webos-image-devel on VMware
3. Verify that HOME application starts well

:QA Notes:
Test-cases: QWTC-10184

:Issues Addressed:
[WRQ-6482] [ose emulator] Using webos-image-devel, blank screen is
 displayed(No home app is displayed) after booting

Change-Id: I13575a26f359393ada0c2d6229565f2f7ed8d1e4
Reviewed-on: http://gpro.lge.com/c/webosose/chromium108/+/390569
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Vladislav Mukulov <vladislav.mukulov@lge.com>
Reviewed-by: Automated Checker <automated_checker@lge.com>
Tested-by: Vladislav Mukulov <vladislav.mukulov@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>