[op][n_upsable][app-shell] Fix the issue with HTTP auth popup not show
:Release Notes:
Fix the issue with HTTP auth popup not show

:Detailed Notes:
This issue is a side effect of Proxy Support Feature
In this feature, we deleted AuthRequestCallback to allow running proxy
authentication directly without entering username and password into
the popup. However, this is wrong. In some cases, AuthRequestCallback
is required from other sections to obtain authentication. To deal with
this problem, I am using AuthChallengeInfo to check where this
authentication request originated so that it can be handled correctly.

:Testing Performed:
Proxy access:
1. Open enactbrowser.
2. Entering website "https://jigsaw.w3.org/HTTP/Basic/"
Popup with *Username* and *password* should be displayed with ok
and cancel button

:QA Notes:
Test-cases: QWTC-11104

:Issues Addressed:
[WRQ-10810][TAS Failed][OSE][QWTC-11104] Basic HTTP auth popup

Change-Id: I2b8a696609c9d614b8db81132182e1ab524071e5
Reviewed-on: http://gpro.lge.com/c/webosose/chromium108/+/393139
Reviewed-by: Automated Checker <automated_checker@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>