[op][n_upsable][ozone] Do not redundantly set screen
:Release Notes:
Do not crash when building with debug checks.

:Detailed Notes:
In m108 Chromium (and likely for some time), each display::Screen
instance created by Ozone backends, is expected to set itself as the
screen instance. Remove in Neva side the code that redundantly set
it.

This avoids hitting a DCHECK

:Testing Performed:
1. Build with dcheck_is_configurable
2. Enable DcheckIsFatal
3. It should not crash on launching

:QA Notes:
Test-cases: N/A

:Issues Addressed:
[WRQ-12785] webruntime crashes in runtime when built with dcheck_always_on=true

Change-Id: Iac083c99006702373beee9a26344bc2076554f36
Reviewed-on: http://gpro.lge.com/c/webosose/chromium108/+/397113
Tested-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>