[op][n_upsable][webrtc] Do not request to drop connections from dead processes
:Release Notes:
Make DCHECK builds usable by fixing reported crashes.

:Detailed Notes:
The API to drop all peer connections, used when suspending a web page,
does not check if the RenderFrameHost process is still alive. That
makes DCHECK fail when trying to obtain the PID of the process.

DropAllPeerConnections should check if the RenderFrameHost is alive.

:Testing Performed:
Build with DCHECK and force a crash of a renderer process. No DCHECK
should fail.

:QA Issues:
Test-cases: N/A

:Issues Addressed:
[WRQ-12785] webruntime crashes in runtime when built with dcheck_always_on=true

Change-Id: Ida5011b8c1a1d26da16e678b044ed002f46b66e7
Reviewed-on: http://gpro.lge.com/c/webosose/chromium108/+/397116
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>